Improved GPS code
Modified the GPS code to read and process incoming data as much as
possible, until an interval occurs. This fixes reliability issues where
the serial port would aparently get a lot of framing errors and dropped
bytes (and therefore, no date, which made the Time library not sync, or
de-sync).

Also added a GPS state to allow the sketch to see if the GPS is
receiving, if it got any valid data, if it got the date and if it got
the time. The higher the state, the more information has been received.